separate session from wppwebjs
