+++ app/wppwebjs/index.ts
@@ -1,3 +1,3 @@
-export * from "./session";
+export * from "../session";
 export * from "./number-utils";
 export * from "./messaging";
